refactor(voice/tests): hoist AudioUserSimulator fixture to fixtures/ (#524)
hooks.test.ts and vad-fallback.test.ts each defined a structurally
identical AudioUserSimulator (same constructor shape, same call()
returning a base64 input_audio/pcm16 content block); vad-fallback also
carried its own audioMessageContent helper. Move both to the existing
fixtures/ directory as fixtures/audio-user-simulator.ts and import from
both test files, dropping the inline definitions and the now-unused
domain imports they pulled in.
